moved localizations to src
