+++ l10n.yaml
@@ -1,8 +1,8 @@
-arb-dir: lib/l10n
-output-dir: lib/l10n
+arb-dir: lib/src/l10n
+output-dir: lib/src/l10n
 template-arb-file: app_en.arb
 use-deferred-loading: true
-gen-inputs-and-outputs-list: lib/l10n
+gen-inputs-and-outputs-list: lib/src/l10n
 synthetic-package: false
 required-resource-attributes: false
 format: true
