+++ lib/src/classes/EzColorSetting.dart
@@ -3,8 +3,7 @@
  * See LICENSE for distribution and usage details.
  */
 
-import '../../l10n/app_localizations.dart';
-
+import '../l10n/app_localizations.dart';
 import '../../empathetech_flutter_ui.dart';
 
 import 'package:flutter/material.dart';
